requirements.txt
asgiref==3.7.2
Django==4.2.7
django-admin-interface==0.26.1
django-colorfield==0.10.1
django-debug-toolbar==4.2.0
django-resized==1.0.2
django-taggit==5.0.1
Pillow==10.1.0
python-slugify==8.0.1
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
